requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic-core==2.14.1
pymongo==4.6.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
gitpython==3.1.40
pygments==2.17.2
coverage==7.3.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
celery==5.3.4
redis==5.0.1
psutil==5.9.6
pathlib2==2.3.7
jinja2==3.1.2
markdown==3.5.1
pyyaml==6.0.1
toml==0.10.2
click==8.1.7
rich==13.7.0
tqdm==4.66.1
